[Refactor][MoE] Merge vLLM correctness test into test_moe_kimi_moe.py
Move test_kimi_moe_vs_vllm into the main test file; use pytest.importorskip
inside the test function so only the vLLM cases are skipped when vLLM is
not installed, without splitting into a separate file.
